requirements.txt
pyyaml
Markdown
MarkupSafe
Jinja2